Created a VideoPlayer with input an assetId Created a backing HlsVideoPlayer
