+++ docs/components/examples/DemoPlayer.tsx
@@ -6,7 +6,8 @@ import { useState } from 'react';
 const assetId: AssetIdOrString = '01c94ad7-35f1-4a55-9802-57c2fa39b964'; // clock
 // const assetId: AssetIdOrString = 'a4e87108-d3b5-4e6b-bb9c-7ca321119cb6'; // waterfall
 
-export const AssetDemoPlayer = () => {
+// TODO: do not merge in production
+export const DemoPlayer = () => {
   const [playbackUrl, setPlaybackUrl] = useState<string | undefined>(undefined);
   const [errorMessage, setErrorMessage] = useState<string | undefined>(
     undefined,
