+++ docs/pages/examples/video-player.en-US.mdx
@@ -3,11 +3,11 @@ title: 'Video Player Example'
 description: 'TODO: just for metrics reporting testing purposes'
 ---
 
-import { AssetDemoPlayer } from '@components/examples/AssetDemoPlayer';
+import { DemoPlayer } from '@components/examples/DemoPlayer';
 
 # Video Player Example
 
 This player is reporting metrics to `wss://patchy.ddvtech.com/mist/json_bunny.js`.
 
 <hr />
-<AssetDemoPlayer />
+<DemoPlayer />
